+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- next-header -->
 ## [Unreleased] - ReleaseDate
+### Changed
+* The MSRV has been updated to 1.81.0 due to `core::error::Error` being implemented
 
 ## [0.2.0] - 2023-11-14
 ### Added